ARM: dts: sun8i-a83t-tbs-a711: Add PN544 NFC support
Regulators on the schematic are named incorrectly. Both cameras are
is in fact connected to the dvdd-csi-f regulator and dvdd-csi-r is
in reality used for digital pad supply for NFC chip.

At the same time vcc-mipi regulator is not used for MIPI, but for NFC
VBAT power.

Interpreting schematics is an art form! :D

Signed-off-by: Ondrej Jirman <megous@megous.com>